server/requirements.txt
tensorflow
pillow
numpy
django
djangorestframework